Fix/support float delta_t

+++ src/fsrs_optimizer/fsrs_optimizer.py
@@ -213,7 +213,7 @@ class BatchDataset(Dataset):
         self.x_train = pad_sequence(
             dataframe["tensor"].to_list(), batch_first=True, padding_value=0
         )
-        self.t_train = torch.tensor(dataframe["delta_t"].values, dtype=torch.int)
+        self.t_train = torch.tensor(dataframe["delta_t"].values, dtype=torch.float)
         self.y_train = torch.tensor(dataframe["y"].values, dtype=torch.float)
         self.seq_len = torch.tensor(
             dataframe["tensor"].map(len).values, dtype=torch.long
